fix(submit-gate): unify quote-quantity admission notional and floor both sides
Finding B (canary path): route the canary proof executor and the strategy
submit path through a single admission_base_notional_from_order so a
price*quantity shortcut can no longer understate a quote-quantity order's
per-order cap. One definition of the admission base notional, used everywhere.

BUY-side floor symmetry: conservative_quote_quantity_admission_notional floored
only SELL; a quote-quantity BUY returned NT's bare effective notional, which
venue base-quantity rounding can leave a sub-tick below the committed quote
quantity (reproduced: BUY 24.9995 vs committed 25.00). The floor now applies to
Buy and Sell alike (non-inverse Limit/StopLimit), making the cap-floor a
structural guarantee instead of a precision coincidence. Inverse instruments
stay skipped.

Also drop a redundant 'as u64' cast flagged by clippy --tests.

Tests: submit-admission 37/37 (incl. 3 new BUY fixtures), strategy
quote-quantity 14/14, lib 420/420, live_canary_gate 77/77; clippy --lib --tests
-D warnings clean, fmt clean.

+++ tests/bolt_v3_live_canary_gate.rs
@@ -904,7 +904,7 @@ fn configure_proof_policy_evidence(operator_evidence: &mut LiveCanaryOperatorEvi
         &std::fs::read(&gate_session_path).expect("gate session should read"),
     )
     .expect("gate session should parse");
-    let now_ms = (current_unix_seconds_for_test() as u64).saturating_mul(1_000);
+    let now_ms = current_unix_seconds_for_test().saturating_mul(1_000);
     gate_session["created_at_ms"] = serde_json::json!(now_ms);
     let gate_session_bytes =
         serde_json::to_vec(&gate_session).expect("gate session should serialize");
